Projects/04_Agentic_Todo_Manager/notification-api/pyproject.toml
[project]
name = "notification-api"
version = "0.1.0"
description = "Notification service for Agentic Todo Manager"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.135.2",
    "psycopg[binary]>=3.2.0",
    "python-dotenv>=1.0.0",
]
